Fixes up capacity script, extricates to lib
